feat(examples): adds stored query example for images
replaces the existing example with the new recommended way to get image sources

refs RBCOM-841
